pdf-mcp/requirements.txt
mcp
google-generativeai
langchain==0.2.7
langchain-community==0.2.6
langchain-google-genai==1.0.7
faiss-cpu
pypdf
python-dotenv
